+++ examples/hello_world.py
@@ -1,13 +1,9 @@
 import shellish
 
-class Hello(shellish.Command):
+class Hello(shellish.InteractiveCommand):
     """ I am a required docstring used to document the --help output! """
 
     name = 'hello'
-
-    def run(self, args):
-        """ Just run the shell if a subcommand was not given. """
-        self.interact()
 
 
 class World(shellish.Command):