+++ examples/skin_a_cat.py
@@ -9,7 +9,8 @@ import shellish
 ##############
 @shellish.autocommand
 def cat1():
-    cat1.interact()
+    """ Auto command cannot interact. """
+    pass
 
 @shellish.autocommand
 def sub1(optional:int=1):
@@ -21,25 +22,22 @@ cat1.add_subcommand(sub1)
 ###############
 # Composition #
 ###############
-cat2 = shellish.Command(name='cat2', title='composition cat')
-cat2.run = lambda args: cat2.interact()
+cat2 = shellish.InteractiveCommand(name='cat2', title='composition cat')
 
 sub2 = shellish.Command(name='sub2', title='composition cat sub')
 sub2.add_argument('--optional', type=int, default=2)
 sub2.run = lambda args: print("ran subcommand2", args.optional)
 cat2.add_subcommand(sub2)
+#cat2()
 
 
 ###############
 # Inheritance #
 ###############
-class Cat3(shellish.Command):
+class Cat3(shellish.InteractiveCommand):
     """ Inheritance cat. """
 
     name = 'cat3'
-
-    def run(self, args):
-        self.interact()
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
@@ -59,8 +57,8 @@ class Sub3(shellish.Command):
 
 
 # Putting it together for a demo..
-main = shellish.Command(name='main', title='harness')
+main = shellish.InteractiveCommand(name='main', title='harness')
 main.add_subcommand(cat1)
 main.add_subcommand(cat2)
 main.add_subcommand(Cat3)
-main.interact()
+main()