+++ shellish/__init__.py
@@ -1,6 +1,6 @@
-from .shell import *
 from .completer import *
 from .command import *
 from .layout import *
 from .data import *
 from .eventing import *
+from .session import *