+++ shellish/command/__init__.py
@@ -1,0 +1,20 @@
+
+import sys
+
+
+def linebuffered_stdout():
+    """ Always line buffer stdout so pipes and redirects are CLI friendly. """
+    if sys.stdout.line_buffering:
+        return sys.stdout
+    orig = sys.stdout
+    new = type(orig)(orig.buffer, encoding=orig.encoding, errors=orig.errors,
+                     line_buffering=True)
+    new.mode = orig.mode
+    return new
+
+sys.stdout = linebuffered_stdout()
+
+from .base import *
+from .autocommand import *
+from .syscomplete import *
+from .interactive import *