+++ test/basic.py
@@ -6,13 +6,6 @@ import shellish
 import unittest
 
 
-class ShellSanity(unittest.TestCase):
-
-    def test_shell_init(self):
-        self.assertRaises(TypeError, shellish.Shell)
-        shellish.Shell(shellish.Command())
-
-
 class CommandSanity(unittest.TestCase):
 
     def test_command_init(self):